client/pyproject.toml
[project]
name = "agentburg-client"
version = "0.1.0"
description = "AgentBurg agent brain client"
requires-python = ">=3.13"
dependencies = [
    "agentburg-shared",
    "litellm>=1.60.0",
    "instructor>=1.7.0",
    "websockets>=14.0",
    "pydantic>=2.10",
    "pydantic-settings>=2.7",
    "pyyaml>=6.0",
    "aiosqlite>=0.20.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=1.0",
    "ruff>=0.9.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agentburg_client"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py313"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "S", "UP"]
ignore = ["S101"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S106", "S108"]
